inference/requirements.txt
numpy==1.24.4
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
transformers==4.51.3
tokenizers==0.21.1
bert-score==0.3.13
peft==0.15.2
safetensors==0.5.3
regex==2024.11.6
huggingface-hub==0.31.1
hf-xet==1.1.0
filelock==3.18.0
fsspec==2025.3.2
torch==2.7.0
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
Pillow==11.2.1
matplotlib==3.9.4
kiwisolver==1.4.7
contourpy==1.3.0
cycler==0.12.1
fonttools==4.58.0
pyparsing==3.2.3
tqdm==4.67.1
gdown==5.1.0
networkx==3.2.1
sympy==1.14.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyYAML==6.0.2
typing_extensions==4.13.2
importlib_resources==6.5.2
requests==2.32.3
charset-normalizer==3.4.2
idna==3.10
urllib3==2.4.0
certifi==2025.4.26
six==1.17.0
zipp==3.21.0
